pyproject.toml
[project]
name = "hnh"
version = "0.1.0"
description = "hotdog"
authors = [
    {name = "rlaehgus97", email = "starkim625@gmail.com"},
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "transformers>=4.45.0",
    "torch>=2.4.1",
    "Pillow>=10.4.0",
    "python-multipart>=0.0.10",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "GNU"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
test = [
    "pytest>=8.3.3",
]
